backend/pyproject.toml
[project]
name = "shiok-lah-backend"
version = "0.1.0"
description = "Shiok Lah! backend — FastAPI + LangGraph + Mistral + ElevenLabs"
requires-python = ">=3.12"

dependencies = [
    # Web framework
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # LLM / agent orchestration
    "langchain-mistralai>=1.1.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    # Database
    "sqlalchemy>=2.0.0",
    # HTTP client (ElevenLabs streaming)
    "httpx>=0.27.0",
    # Env / config
    "python-dotenv>=1.0.0",
    # ElevenLabs SDK (used by voices/create_voices.py)
    "elevenlabs>=1.0.0",
]

[tool.uv]
package = false
